Move CSS utils file
