+++ weasyprint/css/validation/expanders.py
@@ -14,6 +14,8 @@ import functools
 from tinycss2.color3 import parse_color
 
 from ..properties import INITIAL_VALUES, Dimension
+from ..utils import (
+    InvalidValues, get_keyword, get_single_keyword, split_on_comma)
 from .descriptors import expand_font_variant
 from .properties import (
     background_attachment, background_image, background_position,
@@ -22,8 +24,6 @@ from .properties import (
     flex_wrap, font_family, font_size, font_stretch, font_style, font_weight,
     line_height, list_style_image, list_style_position, list_style_type,
     other_colors, overflow_wrap, validate_non_shorthand)
-from .utils import (
-    InvalidValues, get_keyword, get_single_keyword, split_on_comma)
 
 EXPANDERS = {}
